Move function to make it callable from several modules
