+++ src/Module/ModuleOffersReader.php
@@ -160,38 +160,4 @@ class ModuleOffersReader extends ModuleOffers
         $this->Template->offer = $this->parseOffer($this->offer);
         $this->Template->moduleId = $this->id;
     }
-
-    /**
-     * Parse and return an application form for a job.
-     *
-     * @param int    $intId       [Job ID]
-     * @param string $strTemplate [Template name]
-     *
-     * @return string
-     */
-    protected function getApplicationForm($intId, $strTemplate = 'offer_apply')
-    {
-        if (!$this->offer_applicationForm) {
-            return '';
-        }
-
-        $strForm = $this->getForm($this->offer_applicationForm);
-
-        $objItem = OfferModel::findByPk($intId);
-
-        if (!$objItem) {
-            return '';
-        }
-
-        $objTemplate = new FrontendTemplate($strTemplate);
-        $objTemplate->id = $objItem->id;
-        $objTemplate->code = $objItem->code;
-        $objTemplate->title = $objItem->title;
-        $objTemplate->recipient = $GLOBALS['TL_ADMIN_EMAIL'];
-        $objTemplate->time = time();
-        $objTemplate->token = RequestToken::get();
-        $objTemplate->form = $strForm;
-
-        return $objTemplate->parse();
-    }
 }
